DocsQnA/pyproject.toml
[project]
name = "docsqna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.3.0",
    "fastapi>=0.120.4",
    "langchain[openai]>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
    "pymupdf>=1.26.5",
]
